jQuery Conversion and Bug relief
Fixed a lot of bugs, such as jQuery variable re-delcaration (fixed
smooth scroll), and tidied up the code a lot. Need to start looking at
translating the JavaScript for sidenav menu opening to jQuery however.
It's not currently working, but I suspect this is perhaps something to
do with the CSS. May be worthwhile researching jQuery side-menus, not
just JavaScript ones.